Revert "One More Step Towards ORM"
This reverts commit 81d38d70232a84f5cdcd7f2904345fbcc0a1688b.

+++ project_spring_boot_backend/src/main/java/com/app/entities/OrderTableId.java
@@ -24,10 +24,10 @@ public class OrderTableId implements Serializable{
 	private Long orderId; 
 	
 	 @Column(name = "rest_id")
-	 private Long restId;
+	 private Integer restId;
 
 	 @Column(name = "item_id")
-	 private Long itemId;
+	 private Integer itemId;
 	 
 	 
 	 @Override
